requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langgraph
crewai
crewai-tools
langchain_groq
python-dotenv
duckduckgo-search
sentence-transformers